+++ CHANGELOG.md
@@ -3,5 +3,11 @@ All notable changes to the "cryptol" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.5.4]
+### Added
+- Run in cryptol command (ctrl-shift-c)
+- Run selection in Cryptol command (ctrl-shift-e)
+- Get type in Cryptol command (ctrl-shift-t)
+
 ## [Unreleased]
 - Initial release